fbx with materials
Hopefully materials should be uploaded now with the FBX. I've kept the orginals and the textures in place just in case it doesn't work in unity, so you can just drag and drop into Unity instead of an annoying set up.

Hopefully this fixes set up for materials.

After ive commited this, im going to start blocking whats on the trello for tomorrow morning so i can begin the detail modelling. Don't expect me to be waking up early in the morning haha!

Best,
Jayjay (they/them)